requirements.txt
aws-cdk-lib==2.102.0
constructs>=10.0.0,<11.0.0
pyyaml==6.0
pandas==1.5.3
numpy~=1.24.2
python-dateutil~=2.8.2
boto3~=1.26.114
pandas~=1.5.3
aws-cdk.aws-cognito-identitypool-alpha==2.65.0a0
pymongo==4.3.3

# Install Portal requirements
-r portal/requirements.txt
